Move arm device discovery to arm_node.cpp, delete unused arm.cpp
arm.cpp is not built (CMakeLists compiles arm_node.cpp and friends),
so the previous commit's discovery logic was dead code. Move it to
arm_node.cpp where the actual hardcoded /dev/ttyACM0 lived, and
delete arm.cpp entirely.
